requirements.txt
Flask==2.3.2  
Flask-Login==0.5.0  
Flask-SQLAlchemy==2.5.1  
Flask-WTF==1.0.1  
WTForms==3.0.0  
itsdangerous>=2.1.2  
Werkzeug>=2.3.3  
qrcode==7.4.2  
Pillow==10.3.0  
python-dotenv==1.0.1